Move cp_sequence_numbers out of framework. Types go to indexer-alt-schema, and functionality is in a handler in indexer-alt
