+++ crates/sui-indexer-alt/src/handlers/cp_sequence_numbers.rs
@@ -1,25 +1,54 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::schema::cp_sequence_numbers;
+use std::ops::Range;
+use std::sync::Arc;
+
 use anyhow::{bail, Result};
 use diesel::prelude::*;
 use diesel_async::RunQueryDsl;
-use std::ops::Range;
+use sui_indexer_alt_framework::{
+    db,
+    pipeline::{concurrent::Handler, Processor},
+    types::full_checkpoint_content::CheckpointData,
+};
+use sui_indexer_alt_schema::cp_sequence_numbers::StoredCpSequenceNumbers;
+use sui_indexer_alt_schema::schema::cp_sequence_numbers;
 
-use crate::db::Connection;
-use crate::FieldCount;
+pub struct CpSequenceNumbers;
 
-#[derive(Insertable, Selectable, Queryable, Debug, Clone, FieldCount)]
-#[diesel(table_name = cp_sequence_numbers)]
-pub struct StoredCpSequenceNumbers {
-    pub cp_sequence_number: i64,
-    pub tx_lo: i64,
-    pub epoch: i64,
+impl Processor for CpSequenceNumbers {
+    const NAME: &'static str = "cp_sequence_numbers";
+
+    type Value = StoredCpSequenceNumbers;
+
+    fn process(&self, checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
+        let cp_sequence_number = checkpoint.checkpoint_summary.sequence_number as i64;
+        let network_total_transactions =
+            checkpoint.checkpoint_summary.network_total_transactions as i64;
+        let tx_lo = network_total_transactions - checkpoint.transactions.len() as i64;
+        let epoch = checkpoint.checkpoint_summary.epoch as i64;
+        Ok(vec![StoredCpSequenceNumbers {
+            cp_sequence_number,
+            tx_lo,
+            epoch,
+        }])
+    }
+}
+
+#[async_trait::async_trait]
+impl Handler for CpSequenceNumbers {
+    async fn commit(values: &[Self::Value], conn: &mut db::Connection<'_>) -> Result<usize> {
+        Ok(diesel::insert_into(cp_sequence_numbers::table)
+            .values(values)
+            .on_conflict_do_nothing()
+            .execute(conn)
+            .await?)
+    }
 }
 
 /// Inclusive start and exclusive end range of prunable txs.
-pub async fn tx_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
+pub async fn tx_interval(conn: &mut db::Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
     let result = get_range(conn, cps).await?;
 
     Ok(Range {
@@ -30,7 +59,7 @@ pub async fn tx_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Result<R
 
 /// Returns the epochs of the given checkpoint range. `start` is the epoch of the first checkpoint
 /// and `end` is the epoch of the last checkpoint.
-pub async fn epoch_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
+pub async fn epoch_interval(conn: &mut db::Connection<'_>, cps: Range<u64>) -> Result<Range<u64>> {
     let result = get_range(conn, cps).await?;
 
     Ok(Range {
@@ -44,7 +73,7 @@ pub async fn epoch_interval(conn: &mut Connection<'_>, cps: Range<u64>) -> Resul
 /// The values are expected to exist since the cp_sequence_numbers table must have enough information to
 /// encompass the retention of other tables.
 pub(crate) async fn get_range(
-    conn: &mut Connection<'_>,
+    conn: &mut db::Connection<'_>,
     cps: Range<u64>,
 ) -> Result<(StoredCpSequenceNumbers, StoredCpSequenceNumbers)> {
     let Range {
