+++ crates/sui-indexer-alt/src/handlers/tx_affected_addresses.rs
@@ -10,13 +10,14 @@ use diesel_async::RunQueryDsl;
 use itertools::Itertools;
 use sui_indexer_alt_framework::{
     db,
-    models::cp_sequence_numbers::tx_interval,
     pipeline::{concurrent::Handler, Processor},
     types::{full_checkpoint_content::CheckpointData, object::Owner},
 };
 use sui_indexer_alt_schema::{
     schema::tx_affected_addresses, transactions::StoredTxAffectedAddress,
 };
+
+use crate::handlers::cp_sequence_numbers::tx_interval;
 
 pub(crate) struct TxAffectedAddresses;
 
@@ -98,10 +99,11 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        handlers::cp_sequence_numbers::CpSequenceNumbers,
         types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
+
+    use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
     async fn get_all_tx_affected_addresses(conn: &mut db::Connection<'_>) -> Result<Vec<i64>> {
         Ok(tx_affected_addresses::table
