+++ crates/sui-indexer-alt/src/handlers/tx_calls.rs
@@ -9,11 +9,12 @@ use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
     db,
-    models::cp_sequence_numbers::tx_interval,
     pipeline::{concurrent::Handler, Processor},
     types::{full_checkpoint_content::CheckpointData, transaction::TransactionDataAPI},
 };
 use sui_indexer_alt_schema::{schema::tx_calls, transactions::StoredTxCalls};
+
+use crate::handlers::cp_sequence_numbers::tx_interval;
 
 pub(crate) struct TxCalls;
 
@@ -89,11 +90,12 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        handlers::cp_sequence_numbers::CpSequenceNumbers,
         types::{base_types::ObjectID, test_checkpoint_data_builder::TestCheckpointDataBuilder},
         Indexer,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
+
+    use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
     async fn get_all_tx_calls(conn: &mut db::Connection<'_>) -> Result<Vec<StoredTxCalls>> {
         Ok(tx_calls::table
